Moved normal user commands over to MemberCommands cog
