+++ src/discord/utils.py
@@ -1,0 +1,4 @@
+
+
+async def harvest_id(user):
+    return user.replace("<@!", "").replace(">", "")
